requirements.txt
ruamel.yaml==0.15.42
#xxx standard lib email
pymysql
# config?
Jinja2
sqlalchemy
# wxt
python-dotenv
webexteamssdk
